// src/contracts/generated/csi07/ProductEmergencyEntryV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CatalogId } from "./CatalogId";
import type { EmergencyFlowPurposeV1 } from "./EmergencyFlowPurposeV1";
import type { Secret32 } from "./Secret32";

export type ProductEmergencyEntryV1 = { schemaVersion: number, purpose: EmergencyFlowPurposeV1, launchId: string, targetTenantId: string, targetUserId: string | null, resultNonce: Secret32 | null, identityHomeRegionId: CatalogId, expiresAt: string, };